feat: only return UUIDs instead of full records
This service wasn't intended to return *all* the data about pieces, only
the pertinent data for the shortlist route. This isn't a great solution
though because the frontend (Ember) will then have an incomplete piece
record in memory. If it were persisted to the backend we would lose a
lot of information (e.g. the *other* agendaitem versions its linked to).
The simplest solution is to only return UUIDs here and have the frontend
pick up all the necessary data via mu-cl-resources.